sweep: node-size optimizer phase 0+1 harness
- add benchmark.py: phase 0 harness (single-run wall-clock, noise
  floor across seeds, sim-vs-prod calibration incl. per-fleet cal)
- add optimize_catalog.py: phase 1 analytical shape catalog +
  per-def eligibility + theoretical util ceiling, with JSON dump
- simulate.py: expose per-pool workload/ds/raw-alloc breakdowns so
  opt/cal metrics can be derived; snapshot per-bucket arrivals to
  keep simulate() safe to re-run in-process
- instance_specs.py: fill in missing instance sizes (2xl/4xl and
  metal variants) across CPU/GPU families for catalog enumeration
- lint/style cleanups across build_csv, pull_hud, sim_load,
  sim_nodes, sim_report (imports, S311 noqa, formatting)

Notes:
The optimizer separates two utilization numbers: opt (workload over
allocatable+ds, the ranking metric) and cal (used/alloc from
per_pool, matching the prod PromQL/Grafana dashboard). The benchmark
harness measures both plus the seed noise floor so later phases know
what signal-vs-noise threshold their changes must clear. Phase 1 is
purely analytical (no sim runs) — it enumerates every (instance, N)
split per fleet family, filters to shapes that fit each runner def,
and reports the best fit and a uniform-weighted ceiling.

+++ osdc/scripts/node-size-sweep/sim_report.py
@@ -11,7 +11,7 @@ def percentiles(vals: list[float], pcts: list[int]) -> dict[int, float]:
         return {p: float("nan") for p in pcts}
     sv = sorted(vals)
     n = len(sv)
-    return {p: sv[max(0, min(n - 1, int(round(p / 100 * (n - 1)))))] for p in pcts}
+    return {p: sv[max(0, min(n - 1, round(p / 100 * (n - 1))))] for p in pcts}
 
 
 def _weighted_series(per_bucket, used_key: str, alloc_key: str) -> list[float]:
